IMG-218 Separate the fluent implementation and API into separate modules
Created fluent module and moved the implementation classes into that. Extracted interfaces from
implementation classes and placed those in the fluent-api module. Modified tests as required.
