+++ fluent-api/src/main/java/org/codice/imaging/nitf/fluent/NitfSegmentsFlow.java
@@ -14,12 +14,9 @@
  */
 package org.codice.imaging.nitf.fluent;
 
-import java.util.List;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 import org.codice.imaging.nitf.core.DataSource;
-import org.codice.imaging.nitf.core.common.CommonSegment;
 import org.codice.imaging.nitf.core.dataextension.DataExtensionSegment;
 import org.codice.imaging.nitf.core.graphic.GraphicSegment;
 import org.codice.imaging.nitf.core.header.NitfHeader;
@@ -31,22 +28,7 @@ import org.codice.imaging.nitf.core.text.TextSegment;
 /**
  * The NitfSegmentsFlow provides methods for processing the contents of the NITF file.
  */
-public class NitfSegmentsFlow {
-
-    private final Runnable callback;
-
-    private DataSource mDataSource;
-
-    NitfSegmentsFlow(final DataSource dataSource, final Runnable runnable) {
-        if (dataSource == null) {
-            throw new IllegalArgumentException(
-                    "ImageSegmentFlow(): constructor argument 'dataSource' may not be null.");
-        }
-
-        mDataSource = dataSource;
-
-        this.callback = runnable;
-    }
+public interface NitfSegmentsFlow {
 
     /**
      * Passes the NITF header to the supplied consumer.
@@ -54,11 +36,7 @@ public class NitfSegmentsFlow {
      * @param consumer the consumer to pass the NITF header to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow fileHeader(final Consumer<NitfHeader> consumer) {
-        NitfHeader nitfHeader = mDataSource.getNitfHeader();
-        consumer.accept(nitfHeader);
-        return this;
-    }
+    NitfSegmentsFlow fileHeader(Consumer<NitfHeader> consumer);
 
     /**
      * Iterates over the images in the NITF file and passes them to the supplied consumer.
@@ -66,9 +44,7 @@ public class NitfSegmentsFlow {
      * @param consumer The consumer to pass the image segment to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow forEachImageSegment(final Consumer<ImageSegment> consumer) {
-        return forEachSegment(consumer, () -> mDataSource.getImageSegments());
-    }
+    NitfSegmentsFlow forEachImageSegment(Consumer<ImageSegment> consumer);
 
     /**
      * Iterates over the data extension segments in the NITF file and passes them to the supplied consumer.
@@ -76,10 +52,7 @@ public class NitfSegmentsFlow {
      * @param consumer The consumer to pass the data extension segment to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow forEachDataExtensionSegment(
-            final Consumer<DataExtensionSegment> consumer) {
-        return forEachSegment(consumer, () -> mDataSource.getDataExtensionSegments());
-    }
+    NitfSegmentsFlow forEachDataExtensionSegment(Consumer<DataExtensionSegment> consumer);
 
     /**
      * Iterates over the text segments in the NITF file and passes them to the supplied consumer.
@@ -87,9 +60,7 @@ public class NitfSegmentsFlow {
      * @param consumer The consumer to pass the text segment to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow forEachTextSegment(final Consumer<TextSegment> consumer) {
-        return forEachSegment(consumer, () -> mDataSource.getTextSegments());
-    }
+    NitfSegmentsFlow forEachTextSegment(Consumer<TextSegment> consumer);
 
     /**
      * Iterates over the graphic segments in the NITF file and passes them to the supplied consumer.
@@ -97,9 +68,7 @@ public class NitfSegmentsFlow {
      * @param consumer The consumer to pass the graphic segment to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow forEachGraphicSegment(final Consumer<GraphicSegment> consumer) {
-        return forEachSegment(consumer, () -> mDataSource.getGraphicSegments());
-    }
+    NitfSegmentsFlow forEachGraphicSegment(Consumer<GraphicSegment> consumer);
 
     /**
      * Iterates over the label segments in the NITF file and passes them to the supplied consumer.
@@ -107,9 +76,7 @@ public class NitfSegmentsFlow {
      * @param consumer The consumer to pass the label segment to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow forEachLabelSegment(final Consumer<LabelSegment> consumer) {
-        return forEachSegment(consumer, () -> mDataSource.getLabelSegments());
-    }
+    NitfSegmentsFlow forEachLabelSegment(Consumer<LabelSegment> consumer);
 
     /**
      * Iterates over the symbol segments in the NITF file and passes them to the supplied consumer.
@@ -117,9 +84,7 @@ public class NitfSegmentsFlow {
      * @param consumer The consumer to pass the symbol segment to.
      * @return this NitfSegmentsFlow.
      */
-    public final NitfSegmentsFlow forEachSymbolSegment(final Consumer<SymbolSegment> consumer) {
-        return forEachSegment(consumer, () -> mDataSource.getSymbolSegments());
-    }
+    NitfSegmentsFlow forEachSymbolSegment(Consumer<SymbolSegment> consumer);
 
     /**
      * passes this NitfSegmentParserFlow's DataSource to the supplied consumer.
@@ -127,30 +92,10 @@ public class NitfSegmentsFlow {
      * @param dataSourceConsumer the consumer for the data source
      * @return this NitfSegmentsFlow
      */
-    public final NitfSegmentsFlow dataSource(final Consumer<DataSource> dataSourceConsumer) {
-        dataSourceConsumer.accept(mDataSource);
-        return this;
-    }
-
-    private <T extends CommonSegment> NitfSegmentsFlow forEachSegment(final Consumer<T> consumer,
-            final Supplier<List<T>> supplier) {
-        if (consumer != null && supplier != null) {
-            List<T> segments = supplier.get();
-
-            if (segments != null) {
-                for (T segment : segments) {
-                    consumer.accept(segment);
-                }
-            }
-        }
-
-        return this;
-    }
+    NitfSegmentsFlow dataSource(Consumer<DataSource> dataSourceConsumer);
 
     /**
      * Performs necessary cleanup.
      */
-    public final void end() {
-        this.callback.run();
-    }
+    void end();
 }
